requirements.txt
alabaster==1.0.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==23.2.0
babel==2.17.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==2.0.12
click==8.1.7
Django==4.2.9
django-filter==25.1
djangorestframework==3.14.0
docutils==0.21.2
drf-spectacular==0.28.0
fastapi==0.115.6
flake8==7.3.0
flake8-docstrings==1.7.0
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.7
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jose==1.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
mccabe==0.7.0
packaging==24.1
passlib==1.7.4
pillow==11.1.0
pluggy==0.13.1
psycopg2-binary==2.9.10
py==1.11.0
pycodestyle==2.14.0
pydantic==2.10.4
pydantic_core==2.27.2
pydocstyle==6.3.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==6.2.4
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.26.0
roman-numerals-py==3.1.0
rpds-py==0.26.0
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.31
sqlparse==0.5.3
starlette==0.41.3
toml==0.10.2
typing_extensions==4.12.2
uritemplate==4.2.0
urllib3==1.26.19
utils==1.0.2
uvicorn==0.34.0
Werkzeug==3.0.3
